修改下timer的remove算法， 用点内存换些效率。

+++ modules/dbProxy/redisConnect.h
@@ -60,7 +60,7 @@ namespace gsf
 				return reply_->type;
 			}
 
-			int integer() {
+			int64_t integer() {
 				return reply_->integer;
 			}
